Change 'val' to 'test' in test.py
It seems like this is leftover code from YOLOV5. I agree with your decision to change the val file to be called test, as it seems to follow machine learning conventions. However, for the sake of consistency, wouldn't it make sense to point to the test file in the YAML rather than the val file (I'm not sure how you carried things out during training, but if you point to 'val' during validation and during evaluation, this could mean that the results are not representative of how the model performs. 

This is an awesome concept overall though, I really hope the results weren't skewed by this change. Keep up the good work!