Resolve YoloV5 Version Error
It seems like the line here would make a call to an uncaught exception in the current version of YoloV5 (this was fixed in a recent patch). The exception read: AttributeError: 'Upsample' object has no attribute 'recompute_scale_factor'. This solution fixed it, by manually ensuring the blank dict in the stack trace below was not passed.